+++ OpenGl-2D-Renderer-Revamped/components/sprite_sheet.cpp
@@ -387,7 +387,7 @@ unsigned short SpriteSheet::GetSpriteIndexByName(
 	for (size_t i = 0; i < m_UVregionNamesFromConfigFile.size(); i++) {
 
 		if (FastStringCompare(m_UVregionNamesFromConfigFile[i].c_str(), _spriteName)) {
-			return i;
+			return static_cast<unsigned short>(i);
 		}
 	}
